move component to components, handle "no guide found" more cleanly
