+++ airbyte-webapp/src/components/DocumentationPanel/index.tsx
@@ -1,0 +1,1 @@
+export { DocumentationPanel } from "./DocumentationPanel";
