+++ airbyte-webapp/src/locales/en.json
@@ -481,6 +481,7 @@
   "connector.releaseStage.generally_available.description": "<b>Generally Available (GA) connectors</b> have been deemed ready for use in a production environment and is officially supported by Airbyte. Their documentation is considered sufficient to support widespread adoption.",
   "connector.connectorsInDevelopment.docLink": "See our <lnk>documentation</lnk> for more details.",
   "connector.setupGuide": "Setup Guide",
+  "connector.setupGuide.notFound": "No Setup Guide found for this connector.",
 
   "credits.credits": "Credits",
   "credits.whatAreCredits": "What are credits?",
