+++ airbyte-webapp/src/views/Connector/ConnectorDocumentationLayout/ConnectorDocumentationLayout.tsx
@@ -5,9 +5,9 @@ import { FormattedMessage } from "react-intl";
 import { ReflexContainer, ReflexElement, ReflexSplitter } from "react-reflex";
 import styled from "styled-components";
 
+import { DocumentationPanel } from "../../../components/DocumentationPanel/DocumentationPanel";
 import { useDocumentationPanelContext } from "./ConnectorDocumentationContext";
 import styles from "./ConnectorDocumentationLayout.module.css";
-import { DocumentationPanel } from "./DocumentationPanel";
 
 const PanelGrabber = styled.div`
   height: 100vh;
